Feature/aml element content assist (#25)
* narrow down valid elements for params

* use angle brackets to call macros w/ AML params

* create more flexible macro signature lookup

The new mechanism allows to match a macro call against
macro signatures without having to determine whether each of
their parameters is either a normal value parameter or an AML
element parameter. Before, the signature was constructed as a
normalized string representation, requiring to either put double
quotes ("") or angle brackets (<>) into the signature depending
on whether the parameter represented a value or an AML element,
respectively. This resulted in the normalization requiring to
determine whether a parameter was an AML element, and the AML
element determination requiring macro normalization, which led
to infinite recursion.

* make test step proposals for macros w/ AML element params

* make AML element proposals for macros

* use findMacro for macro call validation

* fix macro signature matching with punctuation

* fix variable usage check (twice, code clones are EVIL)